[FIRRTL] Add Annotations to InstanceOp
Add an annotations parameter to InstanceOp that is stored in the
attr-dict.  This has a default value, but not a custom builder with a
default.  Update existing sites that create InstanceOp to pass in a
dummy, empty annotation array.

Use the new custom attr-dict parser/printer for InstanceOp to hide
annotations if they either do not exist or are empty.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>